Rename query-client.js to queryClient.jsx - Update path to queryClient
