+++ src/main.jsx
@@ -7,7 +7,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 
 import App from './App';
 import ColorModeProvider from './utils/ColorModeProvider';
-import { queryClient } from './utils/query-client';
+import { queryClient } from './utils/queryClient';
 import { SnackbarUtilsConfigurator } from './utils/SnackbarUtils';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
